+++ backend/restapi/tests/test_google_api.py
@@ -5,7 +5,7 @@ from rest_framework.test import APIRequestFactory
 from rest_framework.response import Response
 from google.oauth2 import id_token
 from google.auth.transport import requests
-from .models import Employees
+from restapi.models import Employees
 from restapi.verificationAPI import VerifyOAuthTokenApi
 
 class TestVerifyOAuthTokenApi(TestCase):
